chore: change structure to monorepo
