+++ package.json
@@ -1,38 +1,23 @@
 {
-  "name": "proto-deck",
-  "type": "module",
-  "version": "0.0.1",
-  "scripts": {
-    "dev": "astro dev",
-    "start": "npm run deck && astro dev",
-    "build": "astro build",
-    "preview": "astro preview",
-    "astro": "astro",
-    "deck": "tsx src/scripts/deck.ts",
-    "prepare": "husky install",
-    "ci": "rome ci .",
-    "format:write": "rome format --write ."
-  },
-  "dependencies": {
-    "@astrojs/mdx": "^0.19.7",
-    "@astrojs/react": "^2.2.1",
-    "@types/react": "^18.0.21",
-    "@types/react-dom": "^18.0.6",
-    "astro": "^2.9.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
-  },
-  "devDependencies": {
-    "husky": "8.0.3",
-    "lint-staged": "13.2.3",
-    "rome": "12.1.3",
-    "tsx": "3.12.7",
-    "vite-plugin-watch-and-run": "^1.1.3",
-    "vitest": "0.33.0"
-  },
-  "lint-staged": {
-    "*.ts": [
-      "rome format --write"
-    ]
-  }
+	"name": "astro-deck",
+	"type": "module",
+	"version": "0.4.0",
+	"scripts": {
+		"prepare": "husky install",
+		"ci": "rome ci .",
+		"format:write": "rome format --write .",
+		"test": "vitest"
+	},
+	"dependencies": {},
+	"devDependencies": {
+		"husky": "8.0.3",
+		"lint-staged": "13.2.3",
+		"rome": "12.1.3",
+		"tsx": "3.12.7",
+		"vitest": "0.33.0",
+		"typescript": "5.1.6"
+	},
+	"lint-staged": {
+		"*.ts": ["rome format --write"]
+	}
 }
