+++ packages/common/index.ts
@@ -1,0 +1,1 @@
+export { mdxToPresentation } from "./src/mdx-to-presentation";
