+++ packages/common/package.json
@@ -1,0 +1,9 @@
+{
+	"name": "@astro-deck/common",
+	"version": "0.4.0",
+	"description": "Common functionality for all packages",
+	"scripts": {},
+	"keywords": [],
+	"author": "",
+	"license": "ISC"
+}
