+++ packages/common/src/mdx-to-presentation.ts
@@ -1,0 +1,7 @@
+import fs from "node:fs";
+
+export function mdxToPresentation(path: string) {
+	const content = fs.readFileSync(path, "utf-8");
+
+	mdxToPresentation(content);
+}
