+++ packages/wrapper/package.json
@@ -1,0 +1,27 @@
+{
+	"name": "@astro-deck/wrapper",
+	"type": "module",
+	"version": "0.4.0",
+	"description": "The wrapper around Astro",
+	"scripts": {
+		"test": "vitest",
+		"astro": "astro",
+		"dev": "astro dev",
+		"build": "astro build",
+		"preview": "astro preview"
+	},
+	"dependencies": {
+		"@astro-deck/common": "workspace:*",
+		"@astrojs/mdx": "^0.19.7",
+		"@astrojs/react": "^2.2.1",
+		"@types/react": "^18.0.21",
+		"@types/react-dom": "^18.0.6",
+		"astro": "^2.9.0",
+		"react": "^18.0.0",
+		"react-dom": "^18.0.0",
+		"vite-plugin-watch-and-run": "^1.1.3"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC"
+}
