+++ packages/wrapper/src/components/SlideControls.tsx
