+++ packages/wrapper/src/lib/__tests__/get-components-from-slide.spec.ts
