+++ packages/wrapper/src/lib/get-components-from-slide.ts
