+++ packages/wrapper/src/lib/get-slides-from-content.ts
