+++ vitest.workspace.ts
@@ -1,0 +1,1 @@
+export default ["packages/*/vitest.config.{e2e,unit}.ts"];
